+++ Example/HWPanModal/PanController/Indicator/HWTextIndicatorView.h
@@ -1,0 +1,18 @@
+//
+//  HWTextIndicatorView.h
+//  HWPanModal_Example
+//
+//  Created by heath wang on 2019/8/10.
+//  Copyright © 2019 heath wang. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import <HWPanModal/HWPanModalIndicatorProtocol.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface HWTextIndicatorView : UIView <HWPanModalIndicatorProtocol>
+
+@end
+
+NS_ASSUME_NONNULL_END