+++ HWPanModal/Classes/Delegate/HWPanModalPresentationDelegate.h
@@ -18,7 +18,7 @@ NS_ASSUME_NONNULL_BEGIN
  * if user do Screen Pan recognizer, interactive = YES
  */
 @property (nonatomic, assign) BOOL interactive;
-@property (nonatomic, strong, readonly) HWPanModalInteractiveAnimator *interactiveDismissalAnimator;
+@property (nonnull, nonatomic, strong, readonly) HWPanModalInteractiveAnimator *interactiveDismissalAnimator;
 
 @end